+++ Bridge/System/ComponentModel/INotifyPropertyChanged.cs
