+++ Tests/Batch1/Batch1.csproj
@@ -140,6 +140,7 @@
     <Compile Include="ObjectLiteralTests.cs" />
     <Compile Include="Reflection\AssemblyAttributesDefinition.cs" />
     <Compile Include="Reflection\AssemblyTests.cs" />
+    <Compile Include="ComponentModel\BrowsableAttributeTests.cs" />
     <Compile Include="Reflection\AttributeTests.cs" />
     <Compile Include="Reflection\TypeAttributeTests.cs" />
     <Compile Include="Reflection\GetMembersTests.cs" />
